move ai call into correct folder
